+++ src/domain/repositories/bitcoin_flow.rs
@@ -5,8 +5,8 @@ use crate::domain::entities::bitcoin_ledger::{BuyBitcoinDto, SellBitcoinDto};
 #[async_trait::async_trait]
 #[mockall::automock]
 pub trait BitcoinFlowRepository {
-    async fn buy(&self, buy_bitcoin_dto: BuyBitcoinDto) -> Result<i64>;
-    async fn sell(&self, sell_bitcoin_dto: SellBitcoinDto) -> Result<i64>;
+    async fn buy(&self, buy_bitcoin_dto: BuyBitcoinDto) -> Result<i32>;
+    async fn sell(&self, sell_bitcoin_dto: SellBitcoinDto) -> Result<i32>;
     async fn view_all_buy(&self) -> Result<Vec<BuyBitcoinDto>>;
     async fn view_all_sell(&self) -> Result<Vec<SellBitcoinDto>>;
 }
