+++ src/domain/value_objects/spending_scanner.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct SpendingScannerModel {
-    pub id: i64,
+    pub id: i32,
     pub amount: f32,
     pub category: String,
     pub description: String,
